iChange the related_name for HealthFacily->Area
This makes possible to use reverse relations from Areas to
HealthFacilities, improving queries that validates areas as belonging to
facilities. This is useful for etag generation for api calls. The
previous version was hashing a query on all Areas: this was both
inefficient and wrong (a user moved to a different facility would still
have the old etag while the returned list of areas would be different)